requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
numpy
opencv-python
pillow
python-dotenv